scraper/requirements.txt
# Core scraping dependencies
playwright>=1.40.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
lxml>=4.9.0

# Data validation and models
pydantic>=2.5.0
pydantic[email]>=2.5.0

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Web framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Async support
asyncio-mqtt>=0.16.0

# Logging and monitoring
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development tools
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: Airflow integration
apache-airflow>=2.7.0

# Optional: Docker support
docker>=6.1.0
